// src/lib/simbee-schema.ts
export interface components {
    schemas: {
// ...
        ClientConsentLayerDto: {
            id: string;
            client_id: string;
            key: string;
            consent_layer_id?: string;
            /** Format: date-time */
            created_at?: string;
            /** Format: date-time */
            updated_at?: string;
        };
// ...
        ConsentDto: {
            id: string;
            client_user_id: string;
            consent_type: string;
            /** Format: date-time */
            granted_at: string;
            /** Format: date-time */
            created_at?: string;
            /** Format: date-time */
            updated_at?: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}